refactor: moving the implementation into the lib folder
